requirement.txt
# Packages listed in alphabetical order
numpy
g2o-python
opencv-contrib-python
pytest
pyyaml
scikit-learn
matplotlib
bokeh
# pytorch special case, see requirements-pytorch.txt